+++ src/main/java/com/nextroom/oescape/OEscapeApplication.java
@@ -1,0 +1,13 @@
+package com.nextroom.oescape;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OEscapeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(com.nextroom.oescape.OEscapeApplication.class, args);
+	}
+
+}
