+++ src/main/java/com/nextroom/oescape/domain/Hint.java
@@ -1,0 +1,23 @@
+package com.nextroom.oescape.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+
+@Entity
+public class Hint {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "theme_id")
+	private Theme theme;
+	private String code;
+	private String contents;
+	private String answer;
+	private Integer progress;
+}
