+++ src/main/java/com/nextroom/oescape/domain/Shop.java
@@ -1,0 +1,22 @@
+package com.nextroom.oescape.domain;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+
+@Entity
+public class Shop {
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String code;
+	private String name;
+	@OneToMany(mappedBy = "shop", cascade = CascadeType.ALL)
+	private List<Theme> themes = new ArrayList<>();
+}
