+++ src/main/java/com/nextroom/oescape/exceptions/CustomException.java
@@ -1,0 +1,10 @@
+package com.nextroom.oescape.exceptions;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class CustomException extends RuntimeException {
+	private final StatusCode statusCode;
+}
