+++ src/main/java/com/nextroom/oescape/exceptions/StatusCode.java
@@ -1,0 +1,24 @@
+package com.nextroom.oescape.exceptions;
+
+import org.springframework.http.HttpStatus;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum StatusCode {
+
+	/**
+	 * 4xx Bad Request
+	 */
+	BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
+
+	/**
+	 * 2xx OK
+	 */
+	OK(HttpStatus.OK, "성공");
+
+	private final HttpStatus code;
+	private final String message;
+}
