--- src/main/java/com/todayescape/TodayEscapeApplication.java
@@ -1,13 +1,0 @@
-package com.todayescape;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TodayEscapeApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TodayEscapeApplication.class, args);
-    }
-
-}
