--- src/main/java/com/todayescape/domain/Theme.java
@@ -1,29 +1,0 @@
-package com.todayescape.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-
-@Entity
-public class Theme {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "shop_id")
-    private Shop shop;
-    private String title;
-    private Long timeLimit;
-    private Long hintCount;
-    @OneToMany(mappedBy = "theme", cascade = CascadeType.ALL)
-    private List<Hint> hints = new ArrayList<>();
-}
