--- src/main/java/com/todayescape/exceptions/CustomException.java
@@ -1,10 +1,0 @@
-package com.todayescape.exceptions;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException {
-    private final StatusCode statusCode;
-}
