--- src/main/java/com/todayescape/exceptions/StatusCode.java
@@ -1,19 +1,0 @@
-package com.todayescape.exceptions;
-
-import org.springframework.http.HttpStatus;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum StatusCode {
-
-    /**
-     * 4xx Bad Request
-     */
-    BAD_REQUEST(HttpStatus.BAD_REQUEST, "잘못된 요청입니다.");
-
-    private final HttpStatus code;
-    private final String message;
-}
