+++ src/test/java/com/nextroom/oescape/OEscapeApplicationTests.java
@@ -1,0 +1,13 @@
+package com.nextroom.oescape;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class OEscapeApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
