--- src/test/java/com/todayescape/TodayEscapeApplicationTests.java
@@ -1,13 +1,0 @@
-package com.todayescape;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TodayEscapeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
