--- src/test/java/com/todayescape/domain/ThemeTest.java
@@ -1,16 +1,0 @@
-package com.todayescape.domain;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class ThemeTest {
-
-	@Test
-	@DisplayName("테마 생성")
-	void createTheme() {
-		Theme theme = new Theme();
-		assertNotNull(theme);
-	}
-}
